Refactored QueryButton into common components
