+++ frontend/src/metabase/css/query_builder.css
@@ -558,20 +558,6 @@
     font-size: 20pt;
 }
 
-.DataRefererenceQueryButton-circle {
-    border: 1px solid currentColor;
-    border-radius: 99px;
-    width: 1.25rem;
-    height: 1.25rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.DataRefererenceQueryButton-text {
-    max-width: 100%;
-}
-
 .DataReference-paneCount {
     padding-right: 0.6em;
 }
